Alteração das Classes de CRUD de Produto

+++ iSick/grails-app/domain/isick/Produto.groovy
@@ -8,8 +8,8 @@ class Produto {
     String tributacaoPisCofins
     String tributacaoIcms
     Date dataCadastro
-    int precoVenda
-    int precoCusto
+    Double precoVenda
+    Double precoCusto
     String tarja
         
     static constraints = {
